# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: Search and list categories
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived, all]
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
      responses:
        "200":
          description: Categories with current spending summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryCollection"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A category with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category details and spending
      responses:
        "200":
          description: Category totals, recent expenses, and monthly spending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDetail"
        "400":
          description: Invalid category ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A category with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories/{categoryId}/archive:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: archiveCategory
      tags: [categories]
      summary: Archive a category
      responses:
        "200":
          description: Archived category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Currency:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          minLength: 3
          maxLength: 3
        name:
          type: string
          minLength: 1
          maxLength: 80
        symbol:
          type: string
          minLength: 1
          maxLength: 8
        decimalPlaces:
          type: integer
          minimum: 0
          maximum: 4
        isActive:
          type: boolean
      required:
        - id
        - code
        - name
        - symbol
        - decimalPlaces
        - isActive
    CurrencyAmount:
      type: object
      properties:
        currency:
          $ref: "#/components/schemas/Currency"
        total:
          type: string
          pattern: "^[0-9]+(\\.[0-9]{1,2})?$"
        count:
          type: integer
          minimum: 0
        share:
          type: number
          minimum: 0
          maximum: 1
          description: Share of the matching same-currency total. Never compares across currencies.
      required:
        - currency
        - total
        - count
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 80
        slug:
          type: string
          minLength: 1
          maxLength: 100
        parentId:
          type: [string, "null"]
          format: uuid
          description: Null for a top-level category.
        icon:
          type: [string, "null"]
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - parentId
        - icon
        - color
        - status
        - createdAt
        - updatedAt
    CategoryInput:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        slug:
          type: string
          minLength: 1
          maxLength: 100
          description: Derived from the name when omitted.
        parentId:
          type: [string, "null"]
          format: uuid
        icon:
          type: string
          maxLength: 64
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
      required:
        - name
    CategoryUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        slug:
          type: string
          minLength: 1
          maxLength: 100
        parentId:
          type: [string, "null"]
          format: uuid
        icon:
          type: [string, "null"]
          maxLength: 64
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
    CategoryListItem:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          properties:
            totals:
              type: array
              items:
                $ref: "#/components/schemas/CurrencyAmount"
            expenseCount:
              type: integer
              minimum: 0
          required:
            - totals
            - expenseCount
    CategoryCollection:
      type: array
      items:
        $ref: "#/components/schemas/CategoryListItem"
    CategoryRecentExpense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: string
          pattern: "^[0-9]+(\\.[0-9]{1,2})?$"
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        description:
          type: [string, "null"]
      required:
        - id
        - amount
        - date
        - description
    CategoryMonthlySpending:
      type: object
      properties:
        month:
          type: string
          pattern: "^\\d{4}-\\d{2}$"
        totals:
          type: array
          items:
            $ref: "#/components/schemas/CurrencyAmount"
        expenseCount:
          type: integer
          minimum: 0
      required:
        - month
        - totals
        - expenseCount
    CategoryDetail:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/Category"
        parent:
          anyOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
          description: Null for a top-level category.
        children:
          type: array
          description: Subcategories. Always empty for a subcategory.
          items:
            $ref: "#/components/schemas/Category"
        totals:
          type: array
          items:
            $ref: "#/components/schemas/CurrencyAmount"
        expenseCount:
          type: integer
          minimum: 0
        recentExpenses:
          type: array
          items:
            $ref: "#/components/schemas/CategoryRecentExpense"
        monthlySpending:
          type: array
          items:
            $ref: "#/components/schemas/CategoryMonthlySpending"
      required:
        - category
        - parent
        - children
        - totals
        - expenseCount
        - recentExpenses
        - monthlySpending
...